updated the UI, fixed the files and made it easy to handle and made the quiz app working
