--- contact-us.html
@@ -1,96 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Contact / Feedback</title>
-    <link rel="stylesheet" href="./styles.css" />
-  </head>
-  <body>
-    <div id="contact-container">
-      <h1>Contact / Feedback</h1>
-
-      <form id="contact-Form">
-        <label>Name</label>
-        <input type="text" name="name" placeholder="Your name (optional)" />
-
-        <label>Email</label>
-        <input
-          type="email"
-          name="email"
-          placeholder="your@email.com (optional)"
-        />
-
-        <label>Reason for Contact</label>
-        <select id="reason" name="reason" required>
-          <option value="">--Select--</option>
-          <option value="feedback">General Feedback</option>
-          <option value="issue">Report a Quiz Issue</option>
-        </select>
-
-        <!-- Feedback Section -->
-        <div id="feedback-Section" style="display: none">
-          <label>UI Rating (1–5)</label>
-          <input
-            type="number"
-            name="uiRating"
-            min="1"
-            max="5"
-            step="1"
-            placeholder="Rate UI from 1 to 5"
-          />
-
-          <label>App Usability Rating (1–5)</label>
-          <input
-            type="number"
-            name="usabilityRating"
-            min="1"
-            max="5"
-            step="1"
-            placeholder="Rate usability from 1 to 5"
-          />
-
-          <label>Message / Suggestion <span style="color: red">*</span></label>
-          <textarea
-            name="message"
-            required
-            placeholder="Enter your message or suggestion..."
-          ></textarea>
-        </div>
-
-        <!-- Quiz Issue Section -->
-        <div id="issue-Section" style="display: none">
-          <label>Quiz Name <span style="color: red">*</span></label>
-          <input
-            type="text"
-            name="quizName"
-            required
-            placeholder="e.g., HTML Quiz"
-          />
-
-          <label>Question Number <span style="color: red">*</span></label>
-          <input
-            type="number"
-            name="questionNumber"
-            required
-            placeholder="Enter question number"
-          />
-
-          <label>Upload Screenshot (optional)</label>
-          <input type="file" name="screenshot" accept="image/*" />
-
-          <label>Issue Description <span style="color: red">*</span></label>
-          <textarea
-            name="issueDescription"
-            required
-            placeholder="Describe the issue in detail..."
-          ></textarea>
-        </div>
-
-        <button type="submit">Submit</button>
-      </form>
-     
-    </div>
-    <script src="./script.js"></script>
-  </body>
-</html>
