--- cpp-quiz.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>C++ Quiz</title>
-  <link rel="stylesheet" href="styles.css">
-</head>
-<style>
-  .quiz-container a {
-    display: block; /* Ensures proper layout behavior */
-    text-align: right; /* Aligns text content, such as the link, to the right */
-    margin-right: 0; /* Optional if extra spacing is needed on the right */
-    float: right; /* Ensures button stays visually aligned correctly */
-  }
-  
-.question-container {
-      margin-bottom: 20px;
-    }
-    .result {
-      margin-top: 10px;
-      font-weight: bold;
-    }
-  </style>
-<body>
-  <div class="quiz-container">
-    <div class="quiz-section">
-      <h2>C++ Quiz</h2>
-      <div id="cpp-questions"></div>
-    </div>
-    
-    <div id="result"></div>
-    <a href="index.html">Back to Home</a>
-  </div>
-  <script src="script.js"></script>
-</body>
-</html>
