--- dsa-quiz.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="styles.css">
-  <title>DSA-quiz</title>
-  <style>
-    .quiz-container a {
-    float: right; 
-  }
-  
-  /* Style added for the icon's size and spacing */
-  .back-button-icon {
-    width: 32px;
-    height: 32px;
-    margin: 10px 10px 0 0;
-  }
-  </style>
-</head>
-<body>
-  <div class="quiz-container">
-     <a href="index.html" class="back-button-icon" aria-label="Back to Home">
-      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-        <path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"></path>
-        <polyline points="9 22 9 12 15 12 15 22"></polyline>
-      </svg>
-    </a>
-
-    <div class="quiz-section">
-      <h2>DSA Quiz</h2>
-      <div id="dsa-questions"></div>
-    </div>
-    
-    <div id="result"></div>
-    <a href="index.html">Back to Home</a>
-  </div>
-  <script src="script.js"></script>
-</body>
-</html>
