--- profile.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <title>Profile</title>
-  <link rel="stylesheet" href="styles.css" />
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css" />
-</head>
-<body>
-
-  <!-- Navbar -->
-   <header class="mynavbar">
-  <div class="navbar-container">
-    <strong>Quiz Game </strong>
-    <nav class="nav-links">
-      <a class="link" href="index.html" target="_blank">Home</a>
-      <a class="link" href="profile.html" target="_blank">Profile</a>
-      <a class="link" href="scoreboard.html" target="_blank">Scoreboard</a>
-      <a class="link" href="logout.html" target="_blank">Logout</a>
-    </nav>
-  </div>
-</header>
-
-  <!-- Profile Message -->
-  <main style="padding: 4rem 2rem; text-align: center;">
-    <h2 style="font-size: 2rem; color: #444;">No profile page available.</h2>
-    <p style="margin-top: 1rem; font-size: 1.1rem; color: #777;">
-      We're working on building your profile experience. Check back later!
-    </p>
-  </main>
-
-
-  <!-- Footer (optional) -->
-  <footer class="myfooter">
-    <div class="footer-container">
-      <p>&#9679; &nbsp; &nbsp; &copy; <strong>Quiz Game - </strong> Built with <i class="fas fa-heart" style="color: red;"> </i> by Akshada Jadhav</p>
-      <a id="contact us" href="./contact-us.html">Contact Us</a>
-      <div class="footer-links">
-        <a class="link" href="https://github.com/Jadhav124Akshada" target="_blank"><i class="fa-brands fa-github icon"></i>Github</a>
-        <a class="link" href="https://www.linkedin.com/in/akshada-jadhav-4b72062b5?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_medium=android_app" target="_blank"><i class="fa-brands fa-linkedin icon"></i>LinkedIn</a>
-      </div>
-    </div>
-  </footer>
-
-</body>
-</html>
