--- scoreboard.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <title>Scoreboard</title>
-  <link rel="stylesheet" href="styles.css" />
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css" />
-</head>
-<body>
-
-  <!-- Navbar -->
-  <header class="mynavbar">
-  <div class="navbar-container">
-    <strong>Quiz Game </strong>
-    <nav class="nav-links">
-      <a class="link" href="index.html" target="_blank">Home</a>
-      <a class="link" href="profile.html" target="_blank">Profile</a>
-      <a class="link" href="scoreboard.html" target="_blank">Scoreboard</a>
-      <a class="link" href="logout.html" target="_blank">Logout</a>
-    </nav>
-  </div>
-</header>
-
-  <!-- Main Scoreboard Content -->
-  <main class="scoreboard-main">
-    <div class="scoreboard-card">
-      <h2>Scoreboard</h2>
-      <table>
-        <thead>
-          <tr>
-            <th>Score</th>
-            <th>Time</th>
-            <th>Player</th>
-            <th>Date</th>
-          </tr>
-        </thead>
-        <tbody id="scoreboard-data">
-          <!-- JS will populate data here -->
-        </tbody>
-      </table>
-      <a href="index.html" class="back-button">Back to Quiz</a>
-    </div>
-  </main>
-
-</body>
-</html>
